Add OpacityAnimationListener inside OpacityAnimation
